split validation functions from base.py (all working except validate)
